Fixed an issue where netsh trace does not stop
Issue1:
On Win10, netsh is started & stopped right away before it's started again
because of a known issue.

However, when stopping netsh trace right after it started, logman -ets
does not show the session yet and it fails to find a session.

Fix1:
Added a rety logic to Stop-NetshTrace with some delay.

Issue2:
In non-English locale, Stop-NetshTrace fails to parse the ouptut of
logman correctly.

Fix2:
Change how to parse the logman output

Fixed Stop-NetshTrace with non-English locale